requirements.txt
aiohttp
aiosqlite
discord.py
cogwatch
black
python-dotenv
flask